Modify GUI

+++ src/shipz/util/GameEvent.java
@@ -6,14 +6,14 @@ import java.util.EventObject;
  * Created by Paul on 06.05.2016.
  */
 public class GameEvent extends EventObject {
-    private byte _id;
+    private int _id;
 
-    public GameEvent(Object source, byte eventId) {
+    public GameEvent(Object source, int eventId) {
         super(source);
         _id = eventId;
     }
 
-    public byte getId() {
+    public int getId() {
         return _id;
     }
 }
